# Improve performance of `==(::AbstractString, ::AbstractString)`