--- a/base/strings/basic.jl
+++ b/base/strings/basic.jl
@@ -321,7 +321,14 @@ julia> "abc" == "αβγ"
 false
 ```
 """
-==(a::AbstractString, b::AbstractString) = cmp(a, b) == 0
+function ==(a::AbstractString, b::AbstractString)
+    a === b && return true
+    ncodeunits(a) != ncodeunits(b) && return false
+    for (c::AbstractChar, d::AbstractChar) in zip(a, b)
+        c != d && return false
+    end
+    return true
+end
 
 """
     isless(a::AbstractString, b::AbstractString) -> Bool