--- a/test/strings/basic.jl
+++ b/test/strings/basic.jl
@@ -680,6 +680,8 @@ end
 Base.iterate(x::CharStr) = iterate(x.chars)
 Base.iterate(x::CharStr, i::Int) = iterate(x.chars, i)
 Base.lastindex(x::CharStr) = lastindex(x.chars)
+Base.ncodeunits(x::CharStr) = sum(ncodeunits, x.chars)
+
 @testset "cmp without UTF-8 indexing" begin
     # Simple case, with just ANSI Latin 1 characters
     @test "áB" != CharStr("áá") # returns false with bug