Originally I was aiming to optimized `==(::SubString, ::String)` but that was already addressed as part of https://github.com/JuliaLang/julia/pull/35973. Most of the performance increase is probably due to removing allocations.

Current master (594ac899ff)

```julia
julia> @btime invoke(==, Tuple{AbstractString, AbstractString}, $"ABC", $"ABC")
  3.905 ns (0 allocations: 0 bytes)
true

julia> @btime invoke(==, Tuple{AbstractString, AbstractString}, $(SubString("ABC")), $"ABCD")
  23.621 ns (1 allocation: 64 bytes)
false

julia> @btime invoke(==, Tuple{AbstractString, AbstractString}, $(SubString("ABC")), $"ABC")
  24.395 ns (1 allocation: 64 bytes)
true

julia> @btime invoke(==, Tuple{AbstractString, AbstractString}, $(SubString("ABC")), $"DEF")
  14.759 ns (1 allocation: 64 bytes)
false

julia> str = "A" ^ 1000;

julia> @btime invoke(==, Tuple{AbstractString, AbstractString}, $(SubString(str)), $str)
  4.141 μs (1 allocation: 64 bytes)
true
```

Current PR:

```julia
julia> @btime invoke(==, Tuple{AbstractString, AbstractString}, $"ABC", $"ABC")
  4.364 ns (0 allocations: 0 bytes)
true

julia> @btime invoke(==, Tuple{AbstractString, AbstractString}, $(SubString("ABC")), $"ABCD")
  5.533 ns (0 allocations: 0 bytes)
false

julia> @btime invoke(==, Tuple{AbstractString, AbstractString}, $(SubString("ABC")), $"ABC")
  13.041 ns (0 allocations: 0 bytes)
true

julia> @btime invoke(==, Tuple{AbstractString, AbstractString}, $(SubString("ABC")), $"DEF")
  7.283 ns (0 allocations: 0 bytes)
false

julia> str = "A" ^ 1000;

julia> @btime invoke(==, Tuple{AbstractString, AbstractString}, $(SubString(str)), $str)
  2.524 μs (0 allocations: 0 bytes)
true
```