@omus 2020-09-08
Added benchmarks in: https://github.com/JuliaCI/BaseBenchmarks.jl/pull/260