@JeffBezanson 2020-09-08
I'm not sure all strings are required to implement `ncodeunits`. E.g. for `CharStr` it doesn't fully make sense. The numbers are also incomparable if the strings use different encodings.