@omus 2020-09-08
> I'm not sure all strings are required to implement `ncodeunits`

I also wasn't quite sure of that myself. As `zip` requires `length` I should be able to use that safely