@JeffBezanson 2020-09-08
> As `zip` requires `length`

It doesn't, but strings do.