@omus 2020-09-08
> It doesn't, but strings do.

I misread the code. Need to be a little less hasty. I can switch it to `length`