@omus 2020-09-08
As `length` checks that each character `isvalid` we lose most of the performance benefit. I'm looking into using `lastindex` instead.