@omus 2020-09-08
Revised benchmarks using `lastindex` instead of `ncodeunits`:

Current master (594ac89)

```julia
julia> str = String('A':'Z') ^ 100;

julia> @btime invoke(==, Tuple{AbstractString, AbstractString}, $str, $str)
  93.236 ns (0 allocations: 0 bytes)
true

julia> @btime invoke(==, Tuple{AbstractString, AbstractString}, $(SubString(str)), $str)
  11.885 μs (1 allocation: 64 bytes)
true

julia> @btime invoke(==, Tuple{AbstractString, AbstractString}, $(SubString(str)), $(str * 'Z'))
  9.803 μs (1 allocation: 64 bytes)
false

julia> @btime invoke(==, Tuple{AbstractString, AbstractString}, $(SubString(str)), $(reverse(str)))
  15.075 ns (1 allocation: 64 bytes)
false
```

Current PR:

```julia
julia> str = String('A':'Z') ^ 100;

julia> @btime invoke(==, Tuple{AbstractString, AbstractString}, $str, $str)
  4.032 ns (0 allocations: 0 bytes)
true

julia> @btime invoke(==, Tuple{AbstractString, AbstractString}, $(SubString(str)), $str)
  8.229 μs (0 allocations: 0 bytes)
true

julia> @btime invoke(==, Tuple{AbstractString, AbstractString}, $(SubString(str)), $(str * 'Z'))
  12.026 ns (0 allocations: 0 bytes)
false

julia> @btime invoke(==, Tuple{AbstractString, AbstractString}, $(SubString(str)), $(reverse(str)))
  15.002 ns (0 allocations: 0 bytes)
false
```