@JeffBezanson 2020-09-08
Unfortunately I think `lastindex` is also encoding-specific.