@omus 2020-09-08
> Unfortunately I think `lastindex` is also encoding-specific.

Ah yes. I see where was going wrong.

```julia
julia> str = String('A':'Z') ^ 100;

julia> @btime invoke(==, Tuple{AbstractString, AbstractString}, $str, $str)
  4.030 ns (0 allocations: 0 bytes)
true

julia> @btime invoke(==, Tuple{AbstractString, AbstractString}, $(SubString(str)), $str)
  8.518 μs (0 allocations: 0 bytes)
true

julia> @btime invoke(==, Tuple{AbstractString, AbstractString}, $(SubString(str)), $(str * 'Z'))
  8.514 μs (0 allocations: 0 bytes)
false

julia> @btime invoke(==, Tuple{AbstractString, AbstractString}, $(SubString(str)), $(reverse(str)))
  7.535 ns (0 allocations: 0 bytes)
false
```