@omus 2020-09-09
I experimented with using iteration directly to avoid using `Iterators.Stateful` and calling `length`

```julia
function ==(a::AbstractString, b::AbstractString)
    a === b && return true
    next_a = iterate(a)
    next_b = iterate(b)
    while next_a !== nothing && next_b !== nothing
        c, state_a = next_a
        d, state_b = next_b
        c != d && return false
        next_a = iterate(a, state_a)
        next_b = iterate(b, state_b)
    end
    return next_a === nothing && next_b === nothing
end
```

The results have similar performance to what is currently implemented:

```julia
julia> str = String('A':'Z') ^ 100;

julia> @btime invoke(==, Tuple{AbstractString, AbstractString}, $str, $str)
  3.784 ns (0 allocations: 0 bytes)
true

julia> @btime invoke(==, Tuple{AbstractString, AbstractString}, $(SubString(str)), $str)
  8.497 μs (0 allocations: 0 bytes)
true

julia> @btime invoke(==, Tuple{AbstractString, AbstractString}, $(SubString(str)), $(str * 'Z'))
  8.494 μs (0 allocations: 0 bytes)
false

julia> @btime invoke(==, Tuple{AbstractString, AbstractString}, $(SubString(str)), $(reverse(str)))
  8.313 ns (0 allocations: 0 bytes)
false
```
As this code is less readable I'll stay with the current implementation.