@omus 2020-09-09
At this point I don't have any more ideas for ways to tweak this so we just need to decide if we want to go ahead with this change.