@jakobnissen 2023-08-15
Is this issue fixed already with  #47125?