@jakobnissen 2023-08-15
Closing this as fixed in #47125 which uses a similar iterator based approach